requirements.txt
fastapi==0.115.11
uvicorn==0.34.0
python-multipart==0.0.20 
opencv-python-headless==4.11.0.86
numpy==2.2.4
pillow==11.1.0
google-genai==1.5.0
